Updated en_us.json Moved IdealPos calculation back to physics tick Added back chat debug messages
